Introduce a reportable concern to encapsulate reporting criteria
We need to evaluate when an order should be reported to TaxJar in
multiple places. Currently this logic all lives in the reporting
subscriber module and is unfortunately duplicated in both handlers. In a
future commit we want to introduce a Sync or Retry action for failed
transaction syncs, which will also require that we perform these checks.
This change attempts to extract some of that logic in a `Reportable`
concern and provide a `with_reportable` API which takes a block to
evaluate when all conditions are met.

In the next commit we'll attempt to further extract the pre-requisites
for replacing a transaction on TaxJar into this and possibly other
checks specific to that flow.
