+++ lib/super_good/solidus_taxjar.rb
@@ -15,6 +15,7 @@ require "super_good/solidus_taxjar/tax_rate_calculator"
 require "super_good/solidus_taxjar/discount_calculator"
 require "super_good/solidus_taxjar/addresses"
 require "super_good/solidus_taxjar/reporting"
+require "super_good/solidus_taxjar/reportable"
 require "super_good/solidus_taxjar/backfill_transactions"
 
 module SuperGood
